refactor: extract RecordingStatusMessageProvider.swift from RecordingStatusMessageBuilder.swift
Extracts 3 types (RecordingStatusMessageSnapshot struct,
RecordingStatusMessageProvider class, PostTranscriptionStatusPresenter class)
from RecordingStatusMessageBuilder.swift into a new
RecordingStatusMessageProvider.swift.

Byte-identical move; SHA-256 16df175dd349ed141af9e9afa6f952556b8d9b9c41681894db5da80e9dcc1005
verified against origin/main:RecordingStatusMessageBuilder.swift lines 66-162.

Imports: Combine + Foundation.

Reduces RecordingStatusMessageBuilder.swift from 162 to 64 lines. Only the
primary RecordingStatusMessageBuilder enum (the message string generator)
remains in RecordingStatusMessageBuilder.swift.

Tests: 667 (unchanged).
